refactor: extract backends module and evolve to backend= API
Architecture changes:
- Extract MonteCarloSimulation orchestration to simulation.py
- Create backends/ submodule with pluggable execution strategies
  - base.py: ExecutionBackend protocol, make_blocks, worker_run_chunk
  - sequential.py: SequentialBackend (single-threaded)
  - parallel.py: ThreadBackend, ProcessBackend
- Refactor core.py to thin facade with re-exports

API evolution:
- Replace parallel=bool with backend=str parameter
  - Accepts: 'auto', 'sequential', 'thread', 'process', 'torch'
- Add DeprecationWarning for legacy parallel= parameter
- Rename backend.execute() to backend.run() for consistency
- Add supports_batch class attribute for GPU-readiness
- Pre-register 'torch' backend with RuntimeError stub
- Add parallel_backend property as legacy alias
